# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvestFlow Investment Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/me/sessions:
    get:
      operationId: getMySessions
      tags: [users]
      summary: Get active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

  # DASHBOARD
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all available strategies
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, paused, closed]
        - in: query
          name: riskLevel
          schema:
            type: string
            enum: [low, medium, high]
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Strategies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get strategy details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /strategies/{id}/performance:
    get:
      operationId: fetchStrategyPerformance
      tags: [strategies]
      summary: Get strategy performance history
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Performance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformancePoint"

  # POSITIONS (COTAS)
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List my withdrawals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, processing, completed, rejected]
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"

    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      summary: Request withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{id}:
    get:
      operationId: getWithdrawal
      tags: [withdrawals]
      summary: Get withdrawal details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # TRANSACTIONS / EXTRATO
...
  /admin/strategies:
    get:
      operationId: adminListStrategies
      tags: [admin]
      summary: Admin list strategies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"

    post:
      operationId: adminCreateStrategy
      tags: [admin]
      summary: Create strategy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Created strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /admin/strategies/{id}:
    patch:
      operationId: adminUpdateStrategy
      tags: [admin]
      summary: Update strategy
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"

    delete:
      operationId: adminDeleteStrategy
      tags: [admin]
      summary: Delete strategy
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/withdrawals:
    get:
      operationId: adminListWithdrawals
      tags: [admin]
      summary: Admin list all withdrawals
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalPage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        emailVerified:
          type: boolean
        twoFactorEnabled:
          type: boolean
        referralCode:
          type: string
        balance:
          type: number
        totalInvested:
          type: number
        totalYield:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, status, emailVerified, twoFactorEnabled, referralCode, balance, totalInvested, totalYield, createdAt]
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
        confirmPassword:
          type: string
    Session:
      type: object
      properties:
        id:
          type: integer
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
      required: [id, createdAt, lastUsedAt, isCurrent]

    # DASHBOARD SCHEMAS
...
    PerformancePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        value:
          type: number
        yield:
          type: number
        yieldPercentage:
          type: number
      required: [date, value, yield, yieldPercentage]
...
    Strategy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        category:
          type: string
        minInvestment:
          type: number
        totalShares:
          type: integer
        availableShares:
          type: integer
        sharePrice:
          type: number
        aum:
          type: number
        maxDrawdown:
          type: number
        totalReturnPct:
          type: number
        monthlyReturnPct:
          type: number
        status:
          type: string
          enum: [active, paused, closed]
        startDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, riskLevel, category, minInvestment, totalShares, availableShares, sharePrice, aum, maxDrawdown, totalReturnPct, monthlyReturnPct, status, startDate, createdAt]
    StrategyDetail:
      allOf:
        - $ref: "#/components/schemas/Strategy"
        - type: object
          properties:
            performanceHistory:
              type: array
              items:
                $ref: "#/components/schemas/PerformancePoint"
            userPosition:
              $ref: "#/components/schemas/Position"
          required: [performanceHistory]
    StrategyInput:
      type: object
      required: [name, description, riskLevel, category, minInvestment, totalShares, sharePrice]
      properties:
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        category:
          type: string
        minInvestment:
          type: number
        totalShares:
          type: integer
        sharePrice:
          type: number
        maxDrawdown:
          type: number
        startDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, paused, closed]
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        category:
          type: string
        minInvestment:
          type: number
        sharePrice:
          type: number
        maxDrawdown:
          type: number
        status:
          type: string
          enum: [active, paused, closed]
...
    Position:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        strategyId:
          type: integer
        strategyName:
          type: string
        shares:
          type: number
        purchasePrice:
          type: number
        currentValue:
          type: number
        investedAmount:
          type: number
        yieldAmount:
          type: number
        yieldPercentage:
          type: number
        status:
          type: string
          enum: [active, closed]
        purchasedAt:
          type: string
          format: date-time
      required: [id, userId, strategyId, strategyName, shares, purchasePrice, currentValue, investedAmount, yieldAmount, yieldPercentage, status, purchasedAt]
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        method:
          type: string
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        walletAddress:
          type: string
        status:
          type: string
          enum: [pending, approved, processing, completed, rejected]
        transactionHash:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        processedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, method, amount, fee, netAmount, walletAddress, status, createdAt]
    WithdrawalInput:
      type: object
      required: [amount, method, walletAddress]
      properties:
        amount:
          type: number
          minimum: 1
        method:
          type: string
          enum: [pix, usdt_bep20, bitcoin, usdc, bnb]
        walletAddress:
          type: string
...
    WithdrawalPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, totalPages]

    # TRANSACTION SCHEMAS
...